commit 6f46ab2665e264e56997384d484b2f3f42dbc376
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu May 7 18:35:11 2015 -0400

    Update size_overflow hash table

 .../size_overflow_plugin/size_overflow_hash.data   |    1 +
 1 files changed, 1 insertions(+), 0 deletions(-)

commit 09328fcefd1c8954ff1c36d5732681a86bef0593
Merge: 606e87a 90b0ee5
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu May 7 18:23:10 2015 -0400

    Merge branch 'pax-stable2' into grsec-stable2

    Conflicts:
    	fs/open.c

commit 90b0ee5d57d22ec889bb3062d580fb83927ce09a
Merge: 15bf6e6 99e64c4
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu May 7 18:19:18 2015 -0400

    Update to pax-linux-3.14.41-test47.patch:
    - worked around a gcc induced intentional overflow in the bunzip decompressor, reported by Martin Filo (https://bugs.gentoo.org/show_bug.cgi?id=548508)
    - F_SETPIPE_SZ handling ignored pipe_min_size and could trigger the size overflow instrumentation, reported by minipli
    - fixed an integer signedness mixup in a parameter of semop and semtimedop, by minipli

    Merge branch 'linux-3.14.y' into pax-stable2

    Conflicts:
    	arch/arm/include/asm/elf.h
    	fs/binfmt_elf.c

commit 606e87a4d2804bd27ec160869055a011d1066adf
Author: Brad Spengler <spender@grsecurity.net>
Date:   Thu May 7 18:01:36 2015 -0400

    Even though in the history of autoconfig it has never tripped anyone
    up under realistic use cases (e.g. nobody starts a config, enables
    only CONFIG_GRKERNSEC, then exits menuconfig, then loads it back up
    and tries to apply an autoconfig setting) explicitly mention that you
    cannot do it and expect the autoconfig to work.  Due to how the Kbuild
    system works, we can't apply defaults and also allow them to be modified
    via custom settings -- if they're 'select'ed then they can't be modified,
    and using 'select' violates any dependencies that may exist.  Therefore
    we have to resort to using 'default', which after a user has already chosen
    all the settings by virtue of enabling CONFIG_GRKERNSEC and then saving their
    complete kernel config, cannot have any effect as the options have now all
    been chosen and there's no 'default' applicable.

 security/Kconfig |    5 +++++
 1 files changed, 5 insertions(+), 0 deletions(-)

Signature-tree: 9cbb97b530b09e952712663ede7a446e509781a2

+++ ipc/sem.c
@@ -1761,7 +1761,7 @@ static int get_queue_result(struct sem_queue *q)
 }
 
 SYSCALL_DEFINE4(semtimedop, int, semid, struct sembuf __user *, tsops,
-		unsigned, nsops, const struct timespec __user *, timeout)
+		long, nsops, const struct timespec __user *, timeout)
 {
 	int error = -EINVAL;
 	struct sem_array *sma;
@@ -1997,7 +1997,7 @@ out_free:
 }
 
 SYSCALL_DEFINE3(semop, int, semid, struct sembuf __user *, tsops,
-		unsigned, nsops)
+		long, nsops)
 {
 	return sys_semtimedop(semid, tsops, nsops, NULL);
 }
